test(playground): move the media conversion test to its own file
`test_message_helpers.py` is upstream's, and the fork had appended a class plus five
imports to it. Its own file instead, so upstream's is byte-identical again.

That leaves `spanDownloadUtils.test.ts` as the only upstream test file the fork touches,
and only because it genuinely cannot compile otherwise — the media endpoints widened the
generic on `authApiFetch.GET`, changing which overload TypeScript picks for an inline
mock literal.
